#11 Restructured the folder
