+++ d3js/src/learning/CovidChartCountries/data/useGlobalData.js
